% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{centerK}
\alias{centerK}
\title{Centering a kernel matrix}
\usage{
centerK(K)
}
\arguments{
\item{K}{Kernel matrix (class "matrix").}
}
\value{
Centered `K` (class "matrix").
}
\description{
It is equivalent to compute `K` over centered data (i.e. the mean of each
column is subtracted) in Feature Space.
}
\examples{
dat <- matrix(rnorm(250),ncol=50,nrow=5)
K <- Linear(dat)
centerK(K)
}
