% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{centerX}
\alias{centerX}
\title{Centering a squared matrix by row or column}
\usage{
centerX(X, rows = TRUE)
}
\arguments{
\item{X}{Numeric matrix or data.frame of any size.}

\item{rows}{If TRUE, the operation is done by row; otherwise, it is done by
column. (Defaults: TRUE).}
}
\value{
Centered X (class "matrix").
}
\description{
It centers a numeric matrix with dimension \emph{N x N} by row (rows=TRUE) or column
 (rows=FALSE).
}
\examples{
dat <- matrix(rnorm(25),ncol=5,nrow=5)
centerX(dat)
}
