% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{dummy_data}
\alias{dummy_data}
\title{Convert categorical data to dummies.}
\usage{
dummy_data(X, lev = NULL)
}
\arguments{
\item{X}{A matrix, or a data.frame containing factors. (If the columns are of
any other class, they will be coerced into factors anyway).}

\item{lev}{(optional) A vector with the categories ("levels") of each factor.}
}
\value{
X (class: "matrix") after performing one-hot-encoding.
}
\description{
Given a matrix or data.frame containing character/factors, this function
performs one-hot-encoding.
}
\examples{
summary(CO2)
CO2_dummy <- dummy_data(CO2[,1:3],lev=dummy_var(CO2[,1:3]))
CO2_dummy[1:10,1:5]
}
