% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{dummy_var}
\alias{dummy_var}
\title{Levels per factor variable}
\usage{
dummy_var(X)
}
\arguments{
\item{X}{A matrix, or a data.frame containing factors. (If the columns are of
any other class, they will be coerced into factors anyway).}
}
\value{
A list with the levels.
}
\description{
This function gives the categories ("levels") per categorical variable ("factor").
}
\examples{
summary(showdata)
dummy_var(showdata)
}
