% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{frobNorm}
\alias{frobNorm}
\title{Frobenius normalization}
\usage{
frobNorm(X)
}
\arguments{
\item{X}{Numeric matrix of any size. It may be a kernel matrix.}
}
\value{
Frobenius-normalized X (class: "matrix").
}
\description{
This function computes the Frobenius normalization of a matrix.
}
\examples{
dat <- matrix(rnorm(50),ncol=5,nrow=10)
frobNorm(dat)
}
