% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_study.R
\name{KTA}
\alias{KTA}
\title{Kernel-target alignment}
\usage{
KTA(K, y)
}
\arguments{
\item{K}{A kernel matrix (class: "matrix").}

\item{y}{The target variable. A factor with two levels.}
}
\value{
Alignment value.
}
\description{
`KTA()` computes the alignment between a kernel matrix and a target variable.
}
\examples{
K1 <- RBF(iris[1:100,1:4],g=0.1)
y <- factor(iris[1:100,5])
KTA(K1,y)
}
