% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showdata.R
\docType{data}
\name{showdata}
\alias{showdata}
\title{Showdata}
\format{
A data.frame with 100 rows and 5 factor variables:
\describe{
  \item{Favorite.color}{Favorite color}
  \item{Favorite.actress}{Favorite actress}
  \item{Favorite.actor}{Favorite actor}
  \item{Favorite.show}{Favorite show}
  \item{Liked.new.show}{Do you like the new show?}
}
}
\source{
Own
}
\usage{
showdata
}
\description{
A toy dataset that contains the results of a (fictional) survey
commissioned from a well-known streaming platform. The platform invited 100
people to watch footage of their new show before the premiere. After that,
the participants were asked to pick their favorite color, actress, actors
and shows from a list. Finally, they were asked to disclose if they liked the new show.
}
\keyword{datasets}
