% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\docType{data}
\name{soil}
\alias{soil}
\title{Soil microbiota (raw counts)}
\format{
A list containing the following elements:
 \describe{
  \item{abund}{Bacterial abundances of 7396 taxa in 88 sites.}
  \item{metadata}{Samples' metadata}
  \item{taxonomy}{Taxonomic information}
}
}
\usage{
soil
}
\description{
Bacterial abundances in 89 soils from across North and South America.
}
\references{
Lauber CL, Hamady M, Knight R, Fierer N. Pyrosequencing-based
assessment of soil pH as a predictor of soil bacterial community structure at
the continental scale. Appl Environ Microbiol. 2009 Aug;75(15):5111-20.
doi: 10.1128/AEM.00335-09.
}
\keyword{datasets}
