% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerntools-package.R
\docType{package}
\name{kerntools-package}
\alias{kerntools}
\alias{kerntools-package}
\title{kerntools: Kernel Functions and Tools for Machine Learning Applications}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Kernel functions for diverse types of data (including, but not restricted to: nonnegative and real vectors, real matrices, categorical and ordinal variables, sets, strings), plus other utilities like kernel similarity, kernel Principal Components Analysis (PCA) and features' importance for Support Vector Machines (SVMs), which expand other 'R' packages like 'kernlab'.
}
\author{
\strong{Maintainer}: Elies Ramon \email{eramon@everlyrusher.com} (\href{https://orcid.org/0000-0002-7953-8115}{ORCID}) [copyright holder]

}
\keyword{internal}
